Allow an override for the CLUSTER_JOIN
Right now you can specify a cluster join but if you specify a discovery
service it will be replaced by the list received from the discovery
service.

In a swarm, you might want to use swarm's own load balancing as the
cluster address while still want the containers to push their infos
to the service discovery for proxysql to use.